pyproject.toml
[tool.poetry]
authors = ["hemengyang <hmy0119@gmail.com>"]
description = "Smart home."
homepage = "https://hehome.xyz"
license = "MIT"
name = "smart-home"
readme = "README.md"
repository = "https://github.com/he0119/smart-home"
version = "0.7.3"

[tool.poetry.dependencies]
boto3 = "^1.17.32"
celery = "^5.0.0"
django = "^3.2.7"
django-celery-beat = "^2.2.1"
django-cleanup = "^5.2.0"
django-filter = "^2.4.0"
django-graphql-jwt = "^0.3.4"
django-mptt = "^0.13.3"
django-storages = "^1.11.1"
graphene-django = "3.0.0b6"
graphene-file-upload = "^1.3.0"
markdown = "^3.3.3"
pillow = "^8.3.2"
psycopg2 = "^2.8.5"
python = "^3.9"
redis = "^3.5.3"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
django-cors-headers = "^3.4.0"
eventlet = "^0.32.0"
isort = "^5.9.3"
black = "^21.9b0"

[[tool.poetry.source]]
default = true
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.black]
# line-length = 90

[tool.isort]
profile = "black"
# line_length = 90
skip_gitignore = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
